# src/main/openapi/DocstoreSpringSwagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: 'DOCSTORE - API'
  description: "DOCSTORE ECM"
  contact:
    name: C KERVOAZOU
    url: https://github.com/kervoaz
    email: kervoaz@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /definition/{type}:
    post:
      summary: Save a document definition (aka schema or template)
      description: <br>Define rules for a type document. Document of this type must conform to this definition</br>
      operationId: storeDefinition
      tags:
        - Configuration

      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
      requestBody:
        description: the document definition
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSchema'
      responses:
        '201':
          description: Event posted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSchema'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    DocumentSchema:
      type: object
      properties:
        functionalType:
          type: string
          description: type of the document
        allowRevision:
          type: boolean
          description: type of document allows revisions
        mandatoryMetadatas:
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
          description: mandatory metadata checked before save
        storagePathPattern:
          type: string
          description: 'storage path builds according this string. Allowed terms are YEAR, MONTH, DOCTYPE and can be combined /YEAR/MONTH/DOCTYPE. If not specified: YEAR/DOCTYPE'
      required:
        - functionalType
        - allowRevision
    Metadata:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
...
    Error:
      type: object
      required:
        - httpMethod
        - requestUri
        - errors
        - statusCode
        - statusCodeText
        - errorDateTime
      properties:
        httpMethod:
          description: The HTTP request method type
          type: string
          example: POST
        requestUri:
          description: The request URI.
          type: string
          example: https://dcsa.org/dcsa/tnt/v1/events
        errors:
          $ref: '#/components/schemas/Errors'
        statusCode:
          description: The HTTP status code
          type: string
          example: 400
        statusCodeText:
          description: The textual representation of the response status.
          type: string
          example: Bad Request
        errorDateTime:
          description: The date and time (in ISO 8601 format) the error occured.
          type: string
          example: '2019-11-12T07:41:00+08:30'
    Errors:
      type: array
      items:
        type: object
        properties:
          reason:
            description: High level error message.
            type: string
            example: invalidQuery
          message:
            description: Detailed error message.
            type: string
            example: The request did not contain one of the three required query parameters.
        required:
          - reason
          - message
    ####String variables below
...